refactor: apply legacy path, naming to default logic
