+++ funnel-ui/src/pages/Intro/index.tsx
@@ -4,10 +4,16 @@ import { Link } from "react-router-dom";
 
 function Intro() {
   return (
-    <Box>
-      Intro{" "}
-      <Button component={Link} to="/signup/user-info">
-        회원 가입
+    <Box sx={{ display: "flex", gap: 2 }}>
+      <Button
+        variant="contained"
+        component={Link}
+        to="/legacy/signup/user-info"
+      >
+        without use funnel 회원 가입
+      </Button>
+      <Button variant="contained" component={Link} to="/signup/user-info">
+        with use funnel 회원 가입
       </Button>
     </Box>
   );
