+++ funnel-ui/src/routes/index.tsx
@@ -1,18 +1,27 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import Intro from "@/pages/Intro";
-import SignUpPetInfo from "@/pages/SignUpPetInfo";
-import SignUpUserInfo from "@/pages/SignUpUserInfo";
-import SignUpComplete from "@/pages/SignUpComplete";
+import LegacySignUpPetInfo from "@/pages/legacy/SignUpPetInfo";
+import LegacySignUpUserInfo from "@/pages/legacy/SignUpUserInfo";
+import LegacySignUpComplete from "@/pages/legacy/SignUpComplete";
 
 function Routers() {
   return (
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Intro />} />
-        <Route path="/signup/pet-info" element={<SignUpPetInfo />} />
-        <Route path="/signup/user-info" element={<SignUpUserInfo />} />
-        <Route path="/signup/complete" element={<SignUpComplete />} />
+        <Route
+          path="/legacy/signup/pet-info"
+          element={<LegacySignUpPetInfo />}
+        />
+        <Route
+          path="/legacy/signup/user-info"
+          element={<LegacySignUpUserInfo />}
+        />
+        <Route
+          path="/legacy/signup/complete"
+          element={<LegacySignUpComplete />}
+        />
       </Routes>
     </BrowserRouter>
   );
